prequirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
joblib>=1.3.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
